Remove src/error.js file
