--- src/error.js
@@ -1,46 +1,0 @@
-
-const {
-    $TYPE,
-    $LAIRCLIENT,
-}					= require('./constants.js');;
-
-
-class LairClientError extends Error {
-    [$TYPE]				= $LAIRCLIENT;
-    [Symbol.toStringTag]		= LairClientError.name;
-
-    static [Symbol.toPrimitive] ( hint ) {
-	return hint === "number" ? null : `[${this.name} {}]`;
-    }
-
-    constructor ( ...params ) {
-	super( ...params );
-
-	if (Error.captureStackTrace) {
-	    Error.captureStackTrace(this, this.constructor);
-	}
-
-	this.name			= this.constructor.name;
-    }
-
-    [Symbol.toPrimitive] ( hint ) {
-	return hint === "number" ? null : this.toString();
-    }
-
-    toString () {
-	return `[${this.constructor.name}( ${this.message} )]`;
-    }
-}
-class ConversionError extends LairClientError {
-    [Symbol.toStringTag]		= ConversionError.name;
-}
-class ParserError extends LairClientError {
-    [Symbol.toStringTag]		= ParserError.name;
-}
-
-
-module.exports = {
-    LairClientError,
-    ConversionError,
-    ParserError
-};
